refactor(perl-corpus): split parser APIs into srp module (#0000)
